# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phishing Simulation Platform API",
        "description": "Comprehensive API for managing phishing simulation campaigns, employee training, and security awareness programs.",
        "contact": {
            "email": "support@phishingsim.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.2:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/track/{token}/opened": {
            "get": {
                "tags": [
                    "Email Tracking"
                ],
                "summary": "Track email open",
                "description": "Track when an email is opened by returning a 1x1 transparent pixel. This endpoint is called automatically by email clients when images are loaded.",
                "operationId": "d1b87983989d5161115e1369de340669",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Unique campaign token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123def456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tracking pixel returned successfully",
                        "content": {
                            "image/gif": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content - Tracking failed",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid token"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/track/{token}/clicked": {
            "get": {
                "tags": [
                    "Email Tracking"
                ],
                "summary": "Track email link click",
                "description": "Track when a user clicks on a link in a phishing simulation email and redirect to the phishing page",
                "operationId": "5ba6ed01d00cc8a659ea3016cd7fe2bc",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Unique campaign token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123def456"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to phishing page",
                        "headers": {
                            "Location": {
                                "description": "URL to redirect to",
                                "schema": {
                                    "type": "string",
                                    "example": "/api/campaign/abc123def456"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign not found - Redirect to 404 page",
                        "headers": {
                            "Location": {
                                "description": "URL to redirect to",
                                "schema": {
                                    "type": "string",
                                    "example": "https://example.com/404"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid token"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/campaign/{token}": {
            "get": {
                "tags": [
                    "Email Tracking"
                ],
                "summary": "Show phishing simulation page",
                "description": "Display a phishing simulation page based on the campaign template. This page is shown to employees who click on phishing email links.",
                "operationId": "f34f661c61a9685eef0d7e7a126f7fc6",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Unique campaign token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123def456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Phishing page displayed successfully",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "description": "HTML content of the phishing simulation page",
                                    "type": "string",
                                    "example": "<!DOCTYPE html><html><head><title>Password Reset</title></head><body>...</body></html>"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Campaign not found"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid token"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/fake-phishing-page": {
            "get": {
                "tags": [
                    "Email Tracking"
                ],
                "summary": "Show fake phishing page",
                "description": "Display a generic fake phishing page for testing or demonstration purposes. This endpoint is used when a valid campaign token is not available.",
                "operationId": "317a632f755dacaa48dcfa291e9fb0b5",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Optional campaign token",
                        "schema": {
                            "type": "string",
                            "example": "abc123def456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fake phishing page displayed successfully",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "description": "HTML content of the fake phishing page",
                                    "type": "string",
                                    "example": "<!DOCTYPE html><html><head><title>Fake Login</title></head><body>...</body></html>"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page not found - Redirect to 404",
                        "headers": {
                            "Location": {
                                "description": "URL to redirect to",
                                "schema": {
                                    "type": "string",
                                    "example": "https://example.com/404"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token in format (Bearer <token>)"
            }
        }
    },
...
}